Ajout fonctionnalite et recherche
