+++ src/components/ListPrestataire.js
@@ -7,12 +7,39 @@ import { Link } from 'react-router-dom';
 import { accountService } from '../_service/account.service';
 import { jwtDecode } from 'jwt-decode';
 import sem from "../assets/img/seminaire.jpg";
+import "../css/bootstrap.min.css";
+import Form from 'react-bootstrap/Form';
+import { FaSearch } from "react-icons/fa";
+import '../css/style2.css';
 
-const ListPrestataire = () => {
+const ListPrestataire = (props) => {
     const [prestataires, setPrestataires] = useState([]);
     const [clientRole, setClientRole] = useState(true);
     const [isLoggedIn, setIsLoggedIn] = useState(accountService.isAuthenticated());
+    const [query, setQuery] = useState('');
+    const [searchTriggered, setSearchTriggered] = useState(false);
 
+    const handleChange = (event) => {
+        event.preventDefault();
+        setQuery(event.target.value);
+    };
+
+    const filterPrestataires = () => {
+        return prestataires.filter(pres => pres.fonction.toLowerCase().includes(query.toLowerCase()));
+    };
+
+    const performSearch = () => {
+        if (!query.trim()) {
+            alert('Veuillez saisir quelque chose dans le champ de recherche.');
+            return;
+        }
+
+        if (filterPrestataires().length === 0) {
+            alert(`Aucun résultat trouvé pour la fonction : ${query}`);
+        }
+
+        setSearchTriggered(true); 
+    };
     useEffect(() => {
         const isAuthenticated = accountService.isAuthenticated();
         setIsLoggedIn(isAuthenticated);
@@ -44,6 +71,21 @@ const ListPrestataire = () => {
 
     return (
         <div>
+            <center>
+                <div className="display">
+                    <Form.Label className='search-label' htmlFor="inputPassword5">Recherche</Form.Label>
+                    <Form.Control
+                        type="text"
+                        className='search'
+                        placeholder="Recherche de prestataire"
+                        value={query}
+                        onChange={handleChange}
+                    />
+                    <button className="search-btn" onClick={performSearch}>
+                        <FaSearch className='icon-search'/>
+                    </button>
+                </div>
+            </center>
             <h1 style={{ margin: "5rem", textAlign: "center" }}>
                 Nos Prestataires
                 <div>
@@ -57,23 +99,20 @@ const ListPrestataire = () => {
                     )}
                 </div>
             </h1>
-            {prestataires.map(pres => {
-                console.log(pres.image);
-                return (
-                    <RowPrestataire
-                        key={pres.id}
-                        imagep={sem}
-                        altp="pas d'image"
-                        nom={pres.nomEntreprise}
-                        fonction={pres.fonction}
-                        call={pres.telephone}
-                        mail={pres.email}
-                        tarif={pres.tarif}
-                        adresse={pres.adresse}
-                        desc={pres.desEntreprise}
-                    />
-                );
-            })}
+            {searchTriggered && filterPrestataires().map(pres => (
+                <RowPrestataire
+                    key={pres.id}
+                    imagep={sem}
+                    altp="pas d'image"
+                    nom={pres.nomEntreprise}
+                    fonction={pres.fonction}
+                    call={pres.telephone}
+                    mail={pres.email}
+                    tarif={pres.tarif}
+                    adresse={pres.adresse}
+                    desc={pres.desEntreprise}
+                />
+            ))}
         </div>
     );
 };
